SignleAxisPrulasionSystem consist of 2 engines instead of array of pointers to engines

+++ VehicleEquipment/Include/DcEngine.hpp
@@ -7,10 +7,10 @@ class DcEngine : public Engine
 public:
     void setValues(const std::vector<uint8_t>&&) override;
 private:
-    int out1;//TODO to do ctora, piny beda ustawione na poczatku
-    int out2;
-    int pwm;
-    int out1Val;
-    int out2Val;
-    int pwmVal;
+    uint8_t out1;//TODO to do ctora, piny beda ustawione na poczatku
+    uint8_t out2;
+    uint8_t pwm;
+    uint8_t out1Val;
+    uint8_t out2Val;
+    uint8_t pwmVal;
 };
